+++ src/opentype/tables/advanced/lookups/gsub/lookup-type-1.js
@@ -1,4 +1,4 @@
-import { LookupType } from "./gsub.js";
+import { LookupType } from "./gsub-lookup.js";
 
 export class LookupType1 extends LookupType {
   constructor(p) {
